removed non std deps
