--- examples/demo/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "demo"
-version = "1.0.0"
-authors = ["Jagannath E Shahi"]
-edition = "2021"
-
-[dependencies]
-structopt = "0.3"
-config = { version = "0.13.2", default-features = false, features = ["toml", "json"] }
-log = "0.4"
-simplelog = "0.11"
-serde = { version = "1.0.130", features = ["derive", "rc"] }
-serde_json = "1.0"
-
-
-
-[[example]]
-name = "demo"
-path = "examples/demo.rs"
-
